@@ -1,36 +1,38 @@
 {
   "name": "@carnesen/bitcoin-conf",
-  "description": "Options, defaults, and other constants related to the Bitcoin Core software",
+  "description": "Constants and utility methods related to Bitcoin configuration",
   "version": "0.0.2-0",
-  "main": "src/index.js",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "engines": {
+    "node": ">=8.0.0"
+  },
   "scripts": {
     "build": "npm run clean && tsc",
-    "check": "tslint --project .",
+    "build:watch": "npm run build -- --watch",
     "clean": "rimraf lib",
-    "fix": "npm run check -- --fix",
-    "test": "npm run check && npm run unit-test && npm run build",
-    "unit-test": "jest",
-    "watch": "npm run build -- --watch",
+    "lint": "tslint --project .",
+    "lint:fix": "npm run lint -- --fix",
+    "test": "npm run lint && npm run test:unit && npm run build",
+    "test:unit": "jest src",
+    "test:unit:watch": "npm run test:unit -- --watch",
     "preversion": "npm test",
     "postversion": "npm publish"
   },
   "dependencies": {
-    "@types/expand-tilde": "2.0.0",
-    "@types/ini": "1.3.29",
-    "expand-tilde": "2.0.2",
+    "@types/ini": "1.3.30",
     "ini": "1.3.5"
   },
   "devDependencies": {
     "@carnesen/tsconfig": "0.0.1",
-    "@carnesen/tslint-config": "0.0.2",
-    "@types/jest": "23.3.9",
-    "@types/node": "10.12.9",
+    "@carnesen/tslint-config": "0.0.3",
+    "@types/jest": "23.3.10",
+    "@types/node": "10.12.11",
     "jest": "23.6.0",
     "rimraf": "2.6.2",
-    "ts-jest": "23.10.4",
-    "ts-node": "7.0.1",
+    "ts-jest": "23.10.5",
     "tslint": "5.11.0",
-    "typescript": "3.1.6"
+    "typescript": "3.2.1"
   },
   "license": "MIT",
   "author": {
