@@ -1,125 +1,44 @@
-import { isAbsolute, join, resolve } from 'path';
-import { platform } from 'os';
-import { decode } from 'ini';
+import { isAbsolute, join } from 'path';
+import { platform, homedir } from 'os';
 import { readFileSync } from 'fs';
-import expandTilde = require('expand-tilde');
-import { URL } from 'url';
+import { decode } from 'ini';
 
-type Flag = 0 | 1;
+export type Flag = '0' | '1';
+export const isEnabled = (flag?: Flag) => flag === '1';
 
-export type Configuration = Partial<{
+export type BitcoinConf = Partial<{
+  [x: string]: any;
   rpcuser: string;
   rpcpassword: string;
-  rpcauth: string | string[];
-  rpcport: number;
+  rpcport: string;
   rpccookiefile: string;
   testnet: Flag;
   regtest: Flag;
   datadir: string;
-  [propertyName: string]: any;
 }>;
 
-let defaultDatadir: string;
+export let defaultBitcoinDataDir: string;
 switch (platform()) {
   case 'darwin':
-    defaultDatadir = '~/Library/Application Support/Bitcoin/';
+    defaultBitcoinDataDir = join(homedir(), 'Library', 'Application Support', 'Bitcoin');
     break;
   case 'win32':
-    defaultDatadir = resolve(process.env.APPDATA || '', 'Bitcoin');
+    if (!process.env.APPDATA) {
+      throw new Error('Expected to find environment variable "APPDATA"');
+    }
+    defaultBitcoinDataDir = join(process.env.APPDATA, 'Bitcoin');
     break;
   default:
-    defaultDatadir = '~/.bitcoin';
+    defaultBitcoinDataDir = join(homedir(), '.bitcoin');
 }
 
-type LoadConfigurationOptions = {
-  conf?: string; // Absolute or datadir-relative path
-  datadir?: string; // Absolute or relative path
-};
+export const defaultBitcoinConfPath = join(defaultBitcoinDataDir, 'bitcoin.conf');
 
-const loadConfigurationSync = (options: LoadConfigurationOptions = {}) => {
-  const { conf = 'bitcoin.conf', datadir = defaultDatadir } = options;
-  const expandedDatadir = expandTilde(datadir);
-  const expandedConf = expandTilde(conf);
-  let confFilePath: string;
-  if (isAbsolute(expandedConf)) {
-    // In this case datadir is ignored
-    confFilePath = expandedConf;
-  } else {
-    // conf is present as a datadir-relative path
-    confFilePath = join(expandedDatadir, conf);
+export const readBitcoinConfSync = (path: string) => {
+  if (!isAbsolute(path)) {
+    throw new Error(`File path must be absolute`);
   }
-  let confFileContents: string = '';
-  try {
-    confFileContents = readFileSync(confFilePath, { encoding: 'utf8' });
-  } catch (ex) {
-    if (ex.code !== 'ENOENT') {
-      throw ex;
-    }
-  }
-  const loadedConfiguration = decode(confFileContents) as Configuration;
-  return loadedConfiguration;
-};
-
-type LoadRpcHrefOptions = LoadConfigurationOptions;
-export const loadRpcHrefSync = (options: LoadRpcHrefOptions = {}) => {
-  const loadedConfiguration = loadConfigurationSync(options);
-  const url = new URL('http://localhost/');
-  const { rpcuser, rpcpassword } = loadedConfiguration;
-  if (rpcpassword) {
-    // cookie-based auth is disabled
-    if (!rpcuser) {
-      throw new Error('Configuration file had rpcpassword but not rpcuser');
-    }
-    url.username = rpcuser;
-    url.password = rpcpassword;
-  } else {
-    // cookie-based auth is enabled
-    const { rpccookiefile = '.cookie' } = loadedConfiguration;
-    let cookieFilePath: string;
-    if (isAbsolute(rpccookiefile)) {
-      cookieFilePath = rpccookiefile;
-    } else {
-      // rpccookiefile is a relative path
-      let resolvedDatadir: string;
-      if (options.datadir) {
-        // passed datadir takes precedence
-        resolvedDatadir = resolve(expandTilde(options.datadir));
-      } else if (loadedConfiguration.datadir) {
-        // tildes in loaded configuration are not expanded
-        resolvedDatadir = resolve(loadedConfiguration.datadir);
-      } else {
-        resolvedDatadir = resolve(expandTilde(defaultDatadir));
-      }
-      const { regtest, testnet } = loadedConfiguration;
-      let cookieDir: string;
-      if (regtest) {
-        cookieDir = join(resolvedDatadir, 'regtest');
-      } else if (testnet) {
-        cookieDir = join(resolvedDatadir, 'testnet3');
-      } else {
-        cookieDir = resolvedDatadir;
-      }
-      cookieFilePath = join(cookieDir, rpccookiefile);
-      const cookieFileContents = readFileSync(cookieFilePath, { encoding: 'utf8' });
-      const [username, password] = cookieFileContents.split(':');
-      if (!username || !password) {
-        throw new Error('Expected cookie file to contain "username:password"');
-      }
-      url.username = username;
-      url.password = password;
-    }
-  }
-  const { rpcport, regtest, testnet } = loadedConfiguration;
-  let resolvedPort: number;
-  if (rpcport) {
-    resolvedPort = rpcport;
-  } else if (regtest) {
-    resolvedPort = 18443;
-  } else if (testnet) {
-    resolvedPort = 18332;
-  } else {
-    resolvedPort = 8332;
-  }
-  url.port = resolvedPort.toString();
-  return url.href;
+  const fileContents = readFileSync(path, { encoding: 'utf8' });
+  const bitcoinConf: BitcoinConf = decode(fileContents);
+  return bitcoinConf;
 };
